requirements.txt
astronomer-cosmos==1.6.0
apache-airflow-providers-snowflake==6.5.0
snowflake-connector-python==3.13.2
